backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
python-multipart
mysql-connector-python
pymysql
cryptography
passlib
python-jose
bcrypt
openai
pypdf2
python-docx
pillow
numpy
faiss-cpu
requests
# AutoGen dependencies
autogen-agentchat
autogen-ext[openai]
aiofiles
pathlib
typing-extensions
